BackEnd/Supabase/requirements.txt
fastapi==0.135.2
psycopg2-binary==2.9.11
pydantic==2.12.5
python-dotenv==1.2.2
uvicorn==0.40.0